Reduce a bit memory usage by clearing the linecache
This stuff is used when displaying stacktraces. Every modules seems to
be pre-loaded in it. By cleaning it after the application startup, we
free a few 10s of MB and the needed modules (the ones mentioned in
stacktraces) will be loaded when needed.

Added a special analyze_type (linecache) in debug/memory to monitor that
a bit more closely.